+++ tests/test-bootstrap.php
@@ -50,7 +50,18 @@ class BootstrapTest extends WP_UnitTestCase {
 			'base_raw'      => 'https://gist.githubusercontent.com',
 		];
 
-		$actual_org   = (new Bootstrap())->set_repo_type_data([], $org);
+		$actual_org = (new Bootstrap())->set_repo_type_data([], $org);
 		$this->assertEqualSetsWithIndex($expected_org, $actual_org);
 	}
+
+	public function test_get_icon_data() {
+		$icon_data           = ['headers' => [], 'icons'=>[]];
+		$expected['headers'] = ['GistPluginURI' => 'Gist Plugin URI'];
+		$expected['icons']   = ['gist' => 'git-updater-gist/assets/github-logo.svg' ];
+
+		$actual = (new Bootstrap())->set_git_icon_data($icon_data, 'Plugin');
+
+		$this->assertEqualSetsWithIndex($expected, $actual);
+	}
+
 }