+++ CHANGES.md
@@ -1,5 +1,10 @@
 [unreleased]
 
+#### 0.7.0 / 2021-03-15 🎂
+* add filter `gu_post_api_response_body`
+* add filter `gu_get_git_icon_data`
+* more tests added
+
 #### 0.6.0 / 2021-03-13
 * remove constructor
 * update `$auth_required`